Rename aws/core/platform/Time.h to aws/core/platform/AWSTime.h so the SDK compiles fine on case insensitive MacOS which has time.h as a system header.
