+++ aws-cpp-sdk-core/include/aws/core/platform/AWSTime.h
